Implement FIFO queues
